SEBSERV-163 client group template actions
