+++ src/main/java/ch/ethz/seb/sebserver/gui/service/remote/webservice/api/exam/template/DeleteExamTemplate.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright (c) 2021 ETH Zürich, Educational Development and Technology (LET)
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
+package ch.ethz.seb.sebserver.gui.service.remote.webservice.api.exam.template;
+
+import org.springframework.context.annotation.Lazy;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.MediaType;
+import org.springframework.stereotype.Component;
+
+import com.fasterxml.jackson.core.type.TypeReference;
+
+import ch.ethz.seb.sebserver.gbl.api.API;
+import ch.ethz.seb.sebserver.gbl.api.EntityType;
+import ch.ethz.seb.sebserver.gbl.model.EntityProcessingReport;
+import ch.ethz.seb.sebserver.gbl.profile.GuiProfile;
+import ch.ethz.seb.sebserver.gui.service.remote.webservice.api.RestCall;
+
+@Lazy
+@Component
+@GuiProfile
+public class DeleteExamTemplate extends RestCall<EntityProcessingReport> {
+
+    public DeleteExamTemplate() {
+        super(new TypeKey<>(
+                CallType.DELETE,
+                EntityType.EXAM_TEMPLATE,
+                new TypeReference<EntityProcessingReport>() {
+                }),
+                HttpMethod.DELETE,
+                MediaType.APPLICATION_FORM_URLENCODED,
+                API.EXAM_TEMPLATE_ENDPOINT + API.MODEL_ID_VAR_PATH_SEGMENT);
+    }
+
+}
