+++ src/main/java/ch/ethz/seb/sebserver/gui/service/remote/webservice/api/exam/template/GetExamTemplateNames.java
@@ -6,33 +6,37 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package ch.ethz.seb.sebserver.gui.service.remote.webservice.api.exam;
+package ch.ethz.seb.sebserver.gui.service.remote.webservice.api.exam.template;
 
 import java.util.List;
 
 import org.springframework.context.annotation.Lazy;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.MediaType;
 import org.springframework.stereotype.Component;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 
 import ch.ethz.seb.sebserver.gbl.api.API;
 import ch.ethz.seb.sebserver.gbl.api.EntityType;
-import ch.ethz.seb.sebserver.gbl.model.exam.ExamTemplate;
+import ch.ethz.seb.sebserver.gbl.model.EntityName;
 import ch.ethz.seb.sebserver.gbl.profile.GuiProfile;
-import ch.ethz.seb.sebserver.gui.service.remote.webservice.api.PageToListCallAdapter;
+import ch.ethz.seb.sebserver.gui.service.remote.webservice.api.RestCall;
 
 @Lazy
 @Component
 @GuiProfile
-public class GetExamTemplates extends PageToListCallAdapter<ExamTemplate> {
+public class GetExamTemplateNames extends RestCall<List<EntityName>> {
 
-    public GetExamTemplates() {
-        super(
-                GetExamTemplatePage.class,
+    public GetExamTemplateNames() {
+        super(new TypeKey<>(
+                CallType.GET_NAMES,
                 EntityType.EXAM_TEMPLATE,
-                new TypeReference<List<ExamTemplate>>() {
-                },
-                API.EXAM_TEMPLATE_ENDPOINT);
+                new TypeReference<List<EntityName>>() {
+                }),
+                HttpMethod.GET,
+                MediaType.APPLICATION_FORM_URLENCODED,
+                API.EXAM_TEMPLATE_ENDPOINT + API.NAMES_PATH_SEGMENT);
     }
 
 }
