+++ src/main/java/ch/ethz/seb/sebserver/gui/service/remote/webservice/api/exam/template/GetExamTemplates.java
@@ -6,11 +6,11 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package ch.ethz.seb.sebserver.gui.service.remote.webservice.api.exam;
+package ch.ethz.seb.sebserver.gui.service.remote.webservice.api.exam.template;
+
+import java.util.List;
 
 import org.springframework.context.annotation.Lazy;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.MediaType;
 import org.springframework.stereotype.Component;
 
 import com.fasterxml.jackson.core.type.TypeReference;
@@ -19,21 +19,19 @@ import ch.ethz.seb.sebserver.gbl.api.API;
 import ch.ethz.seb.sebserver.gbl.api.EntityType;
 import ch.ethz.seb.sebserver.gbl.model.exam.ExamTemplate;
 import ch.ethz.seb.sebserver.gbl.profile.GuiProfile;
-import ch.ethz.seb.sebserver.gui.service.remote.webservice.api.RestCall;
+import ch.ethz.seb.sebserver.gui.service.remote.webservice.api.PageToListCallAdapter;
 
 @Lazy
 @Component
 @GuiProfile
-public class NewExamTemplate extends RestCall<ExamTemplate> {
+public class GetExamTemplates extends PageToListCallAdapter<ExamTemplate> {
 
-    public NewExamTemplate() {
-        super(new TypeKey<>(
-                CallType.NEW,
+    public GetExamTemplates() {
+        super(
+                GetExamTemplatePage.class,
                 EntityType.EXAM_TEMPLATE,
-                new TypeReference<ExamTemplate>() {
-                }),
-                HttpMethod.POST,
-                MediaType.APPLICATION_FORM_URLENCODED,
+                new TypeReference<List<ExamTemplate>>() {
+                },
                 API.EXAM_TEMPLATE_ENDPOINT);
     }
